Fixed errors related with new OTAPI

+++ TUI/Base/VisualContainer.cs
@@ -157,7 +157,7 @@ namespace TerrariaUI.Base
         protected void InheritParentStyle()
         {
             VisualObject node = this;
-            byte? wall = null;
+            ushort? wall = null;
             byte? wallColor = null;
             while (node != null)
             {
